+++ lib/duckdb.rb
@@ -4,6 +4,7 @@ require 'duckdb/database'
 require 'duckdb/connection'
 require 'duckdb/result'
 require 'duckdb/prepared_statement'
+require 'duckdb/appender'
 
 # DuckDB provides Ruby interface of DuckDB.
 module DuckDB